+++ lib/src/core/google_map.dart
@@ -2,19 +2,14 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:flutter/widgets.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
+import 'package:google_directions_api/google_directions_api.dart' show GeoCoord, DirectionsService;
 
-import 'package:google_directions_api/google_directions_api.dart'
-    show GeoCoord, DirectionsService;
-
+import 'google_map.state.dart' if (dart.library.html) '../web/google_map.state.dart' if (dart.library.io) '../mobile/google_map.state.dart';
 import 'map_items.dart';
 import 'map_operations.dart';
 import 'map_preferences.dart';
-
-import 'google_map.state.dart'
-    if (dart.library.html) '../web/google_map.state.dart'
-    if (dart.library.io) '../mobile/google_map.state.dart';
 
 /// This widget will try to occupy all available space
 class GoogleMap extends StatefulWidget {
@@ -33,13 +28,7 @@ class GoogleMap extends StatefulWidget {
     this.initialPosition = const GeoCoord(_defaultLat, _defaultLng),
     this.mobilePreferences = const MobileMapPreferences(),
     this.webPreferences = const WebMapPreferences(),
-  })  : assert(mapType != null),
-        assert(interactive != null),
-        assert(initialPosition != null),
-        assert(initialZoom != null),
-        assert(mobilePreferences != null),
-        assert(webPreferences != null),
-        super(key: key);
+  }) : super(key: key);
 
   /// The initial position of the map's camera.
   final GeoCoord initialPosition;
@@ -97,8 +86,7 @@ class GoogleMap extends StatefulWidget {
 
   /// Gets [MapOperations] interface via provided `key` of
   /// [GoogleMapStateBase] state.
-  static MapOperations? of(GlobalKey<GoogleMapStateBase> key) =>
-      key.currentState;
+  static MapOperations? of(GlobalKey<GoogleMapStateBase> key) => key.currentState;
 
   /// Initializer of [GoogleMap].
   ///
@@ -110,11 +98,8 @@ class GoogleMap extends StatefulWidget {
   GoogleMapState createState() => GoogleMapState();
 }
 
-abstract class GoogleMapStateBase extends State<GoogleMap>
-    implements MapOperations {
+abstract class GoogleMapStateBase extends State<GoogleMap> implements MapOperations {
   @protected
   String fixAssetPath(String icon) =>
-      icon.endsWith('/marker_a.png') || icon.endsWith('/marker_b.png')
-          ? 'packages/flutter_google_maps/'
-          : '';
+      icon.endsWith('/marker_a.png') || icon.endsWith('/marker_b.png') ? 'packages/flutter_google_maps/' : '';
 }